fix(widgetbook): split move army invasion intel catalog part (#4216)
Extract DLG20001 invasion-intel Widgetbook fixtures and use cases into
catalog_dialogs_move_army_invasion_intel.dart so catalog_dialogs.dart
stays under the 1000 non-comment line repo limit.

Refs #4216
